+++ packages/backend-deployer/src/cdk_error_mapper.ts
@@ -99,7 +99,7 @@ export class CdkErrorMapper {
   }> => [
     {
       errorRegex:
-        /ExpiredToken: .*|(Error|InvalidClientTokenId): The security token included in the request is (expired|invalid)/,
+        /ExpiredToken: .*|The security token included in the request is (expired|invalid)/,
       humanReadableErrorMessage:
         'The security token included in the request is invalid.',
       resolutionMessage:
@@ -162,6 +162,16 @@ export class CdkErrorMapper {
       resolutionMessage:
         "Ensure dependencies in your project are installed with your package manager. For example, by running 'yarn install' or 'npm install'",
       errorName: 'CDKNotFoundError',
+      classification: 'ERROR',
+    },
+    {
+      errorRegex:
+        /ValidationError: Role (?<roleArn>.*) is invalid or cannot be assumed/,
+      humanReadableErrorMessage:
+        'Role {roleArn} is invalid or cannot be assumed',
+      resolutionMessage:
+        'Ensure the role exists and AWS credentials have an IAM policy that grants sts:AssumeRole for the role',
+      errorName: 'InvalidOrCannotAssumeRoleError',
       classification: 'ERROR',
     },
     {
@@ -477,6 +487,32 @@ export class CdkErrorMapper {
       classification: 'ERROR',
     },
     {
+      // This error pattern is observed when circular dependency is between stacks but not resources in a stack
+      errorRegex:
+        /ValidationError: Circular dependency between resources: \[(?<resources>.*)\]/,
+      humanReadableErrorMessage:
+        'The CloudFormation deployment failed due to circular dependency found between nested stacks [{resources}]',
+      resolutionMessage: `If you are using functions then you can assign them to existing nested stacks that are dependent on functions or functions depend on them, for example:
+1. If your function is defined as auth triggers, you should assign this function to auth stack.
+2. If your function is used as data resolver or calls data API, you should assign this function to data stack.
+To assign a function to a different stack, use the property 'resourceGroupName' in the defineFunction call and choose auth, data or any custom stack.
+
+If your circular dependency issue is not resolved with this workaround, please create an issue here https://github.com/aws-amplify/amplify-backend/issues/new/choose
+`,
+      errorName: 'CloudformationStackCircularDependencyError',
+      classification: 'ERROR',
+    },
+    {
+      // This error pattern is observed when circular dependency is between resources in a single stack, i.e. ValidationError is absent from the error message
+      errorRegex:
+        /(?<!ValidationError: )Circular dependency between resources: \[(?<resources>.*)\]/,
+      humanReadableErrorMessage:
+        'The CloudFormation deployment failed due to circular dependency found between resources [{resources}] in a single stack',
+      resolutionMessage: `If you are creating custom stacks or adding new CDK resources to amplify stacks, ensure that there are no cyclic dependencies. For more details see: https://aws.amazon.com/blogs/infrastructure-and-automation/handling-circular-dependency-errors-in-aws-cloudformation/`,
+      errorName: 'CloudformationResourceCircularDependencyError',
+      classification: 'ERROR',
+    },
+    {
       errorRegex:
         /(?<stackName>amplify-[a-z0-9-]+)(.*) failed: ValidationError: Stack:(.*) is in (?<state>.*) state and can not be updated/,
       humanReadableErrorMessage:
@@ -513,6 +549,8 @@ export type CDKDeploymentError =
   | 'CDKResolveAWSAccountError'
   | 'CDKVersionMismatchError'
   | 'CFNUpdateNotSupportedError'
+  | 'CloudformationResourceCircularDependencyError'
+  | 'CloudformationStackCircularDependencyError'
   | 'CloudFormationDeletionError'
   | 'CloudFormationDeploymentError'
   | 'CommonNPMError'
@@ -523,6 +561,7 @@ export type CDKDeploymentError =
   | 'ExpiredTokenError'
   | 'FileConventionError'
   | 'ModuleNotFoundError'
+  | 'InvalidOrCannotAssumeRoleError'
   | 'InvalidPackageJsonError'
   | 'SecretNotSetError'
   | 'SyntaxError'