+++ packages/backend/src/engine/backend_id_scoped_ssm_environment_entries_generator.ts
@@ -1,11 +1,13 @@
-import { ParameterPathConversions } from '@aws-amplify/platform-core';
+import {
+  NamingConverter,
+  ParameterPathConversions,
+} from '@aws-amplify/platform-core';
 import {
   BackendIdentifier,
   SsmEnvironmentEntriesGenerator,
 } from '@aws-amplify/plugin-types';
 import { StringParameter } from 'aws-cdk-lib/aws-ssm';
 import { Construct } from 'constructs';
-import { toScreamingSnakeCase } from './naming_convention_conversions.js';
 
 /**
  * Generates SsmEnvironmentEntry[] with SSM parameters that are scoped to a specific backend identifier
@@ -50,7 +52,9 @@ export class BackendIdScopedSsmEnvironmentEntriesGenerator
    */
   generateSsmEnvironmentEntries = (scopeContext: Record<string, string>) =>
     Object.entries(scopeContext).map(([contextKey, contextValue]) => {
-      const sanitizedContextKey = toScreamingSnakeCase(contextKey);
+      const sanitizedContextKey = new NamingConverter().toScreamingSnakeCase(
+        contextKey
+      );
       const parameterPath =
         ParameterPathConversions.toResourceReferenceFullPath(
           this.backendId,